fix: parse float zone coordinates from ZM 1.37+
ZoneMinder changed zone coordinates from pixel integers to percentage
floats. _parse_zone_coords now uses float() instead of int().

Ref: pliablepixels/zmeventnotification#18

+++ pyzm/models/zm.py
@@ -78,7 +78,7 @@ class PTZCapabilities:
 class Zone:
     """A detection zone polygon belonging to a monitor."""
     name: str
-    points: list[tuple[int, int]]
+    points: list[tuple[float, float]]
     pattern: str | None = None
     ignore_pattern: str | None = None
